feat(peagen): move schemas to jsonschemas
