+++ pkgs/standards/peagen/docs/pea-ips/ip-0003.md
@@ -29,7 +29,7 @@ Recent schema linting uncovered several inconsistencies that let malformed input
 
 ### 3.1  Add `$comment` to every schema node
 
-* For **all** schema files under `peagen/schemas/**` add human-readable explanations via `$comment`.
+* For **all** schema files under `peagen/jsonschemas/**` add human-readable explanations via `$comment`.
 * Comments MUST explain *why* a constraint exists, not just *what* it does.
 
 ### 3.2  Constrain `peagen_version`
