+++ pkgs/standards/peagen/peagen/core/process_core.py
@@ -22,7 +22,7 @@ from peagen._utils._search_template_sets import (
     build_file_template_search_paths,
 )
 from peagen.core.validate_core import _collect_errors
-from peagen.schemas import PROJECTS_PAYLOAD_V1_SCHEMA
+from peagen.jsonschemas import PROJECTS_PAYLOAD_V1_SCHEMA
 from peagen.errors import (
     ProjectsPayloadValidationError,
     ProjectsPayloadFormatError,
