+++ pkgs/standards/peagen/peagen/jsonschemas/doe_spec.schema.v1.json
@@ -1,0 +1,71 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/doe_spec.schema.json",
+  "title": "DOE Spec Schema (v1.0.1)",
+  "schemaVersion": "1.0.1",
+  "type": "object",
+
+  "properties": {
+    "schemaVersion": {
+      "type": "string",
+      "const": "1.0.1",
+      "description": "DOE spec schema version"
+    },
+
+    "FACTORS": {
+      "type": "object",
+      "patternProperties": {
+        "^[A-Za-z0-9_]+$": {
+          "type": "object",
+          "properties": {
+            "code":        { "type": "string" },
+            "description": { "type": "string" },
+            "type":        { "type": "string" },
+
+            "levels": {
+              "type": "array",
+              "items": {
+                "oneOf": [
+                  {
+                    "type": "object",
+                    "properties": {
+                      "name": { "type": "string" },
+                      "requirements": {
+                        "type": "array",
+                        "items": { "type": "string" }
+                      }
+                    },
+                    "required": ["name", "requirements"],
+                    "additionalProperties": true
+                  },
+                  { "type": ["string", "number", "boolean", "null"] }
+                ]
+              }
+            },
+
+            "patches": {
+              "type": "array",
+              "items": {
+                "type": "object",
+                "properties": {
+                  "op":   { "type": "string" },
+                  "path": { "type": "string" },
+                  "value": {},
+                  "when": {}
+                },
+                "required": ["op", "path"],
+                "additionalProperties": false
+              }
+            }
+          },
+          "required": ["code", "description", "levels", "patches"],
+          "additionalProperties": true
+        }
+      },
+      "additionalProperties": false
+    }
+  },
+
+  "required": ["schemaVersion", "FACTORS"],
+  "additionalProperties": false
+}
