+++ pkgs/standards/peagen/peagen/jsonschemas/eval_manifest.schema.v1.json
@@ -1,0 +1,76 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://peagen.dev/schemas/eval_manifest-v1.json",
+  "title": "Peagen Evaluation Manifest (v1)",
+  "schemaVersion": "1.0.0",
+  "type": "object",
+  "required": [
+    "schemaVersion",
+    "peagen_version",
+    "workspace_slug",
+    "generated_at",
+    "pool",
+    "evaluators",
+    "results"
+  ],
+  "properties": {
+    "schemaVersion": {
+      "const": "1.0.0",
+      "description": "Hard-coded version tag; bump only on breaking changes."
+    },
+    "peagen_version": {
+      "type": "string",
+      "pattern": "^(?:0|[1-9]\\d*)\\.(?:0|[1-9]\\d*)\\.(?:0|[1-9]\\d*)$",
+      "description": "Semantic version of the Peagen CLI that generated this manifest."
+    },
+    "workspace_slug": {
+      "type": "string",
+      "description": "Slug generated by `peagen program fetch`."
+    },
+    "generated_at": {
+      "type": "string",
+      "format": "date-time",
+      "description": "UTC timestamp (ISO-8601) when evaluation completed."
+    },
+    "pool": {
+      "type": "string",
+      "description": "Fully-qualified class or entry-point of the pool used."
+    },
+    "evaluators": {
+      "type": "array",
+      "description": "Ordered list of evaluator identifiers.",
+      "items": { "type": "string" }
+    },
+    "results": {
+      "type": "array",
+      "minItems": 1,
+      "items": {
+        "type": "object",
+        "required": ["program_path", "scores"],
+        "properties": {
+          "program_path": {
+            "type": "string",
+            "description": "Relative path from workspace root."
+          },
+          "scores": {
+            "type": "object",
+            "description": "Evaluator scores keyed by evaluator id.",
+            "additionalProperties": {
+              "anyOf": [
+                {"type": "number", "minimum": 0, "maximum": 1},
+                {"type": "null"}
+              ]
+            }
+          },
+          "metadata": {
+            "type": "object",
+            "description": "Arbitrary per-evaluator details — nested objects allowed.",
+            "additionalProperties": true
+          }
+        },
+        "additionalProperties": false
+      }
+    }
+  },
+  "additionalProperties": false
+}
