+++ pkgs/standards/peagen/peagen/jsonschemas/extras/cpp_python_pkg.schema.v1.json
@@ -1,0 +1,14 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/extras/cpp_python_pkg.schema.json",
+  "title": "EXTRAS Schema for cpp_python_pkg",
+  "type": "object",
+  "properties": {
+    "AUTHORS": {},
+    "PURPOSE": {},
+    "DESCRIPTION": {},
+    "REQUIREMENTS": {},
+    "DEPENDENCIES": {}
+  },
+  "additionalProperties": false
+}
