+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_base.schema.v1.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/extras/swarmauri_base.schema.json",
+  "title": "EXTRAS Schema for swarmauri_base",
+  "type": "object",
+  "properties": {
+    "RESOURCE_KIND": {},
+    "PURPOSE": {},
+    "DESCRIPTION": {},
+    "REQUIREMENTS": {},
+    "DEPENDENCIES": {},
+    "INTERFACE_FILE": {},
+    "EXTERNAL_DOC_FILES": {},
+    "MIXIN_FILES": {},
+    "INTERFACE_NAME": {},
+    "EXAMPLES": {}
+  },
+  "additionalProperties": false
+}
