+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_core.schema.v1.json
@@ -1,0 +1,18 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/extras/swarmauri_core.schema.json",
+  "title": "EXTRAS Schema for swarmauri_core",
+  "type": "object",
+  "properties": {
+    "RESOURCE_KIND": {},
+    "PURPOSE": {},
+    "DESCRIPTION": {},
+    "REQUIREMENTS": {},
+    "DEPENDENCIES": {},
+    "BASE_FILE": {},
+    "EXTERNAL_DOC_FILES": {},
+    "MIXIN_FILES": {},
+    "EXAMPLES": {}
+  },
+  "additionalProperties": false
+}
