+++ pkgs/standards/peagen/peagen/jsonschemas/extras/swarmauri_standard_standalone.schema.v1.json
@@ -1,0 +1,21 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/extras/swarmauri_standard_standalone.schema.json",
+  "title": "EXTRAS Schema for swarmauri_standard_standalone",
+  "type": "object",
+  "properties": {
+    "RESOURCE_KIND": {},
+    "PURPOSE": {},
+    "DESCRIPTION": {},
+    "REQUIREMENTS": {},
+    "DEPENDENCIES": {},
+    "AUTHORS": {},
+    "BASE_FILE": {},
+    "EXTERNAL_DOC_FILES": {},
+    "MIXIN_FILES": {},
+    "BASE_CLASS": {},
+    "MIXINS": {},
+    "EXAMPLES": {}
+  },
+  "additionalProperties": false
+}
