+++ pkgs/standards/peagen/peagen/jsonschemas/extras/test3.schema.v1.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/extras/test3.schema.json",
+  "title": "EXTRAS Schema for test3",
+  "type": "object",
+  "properties": {},
+  "additionalProperties": false
+}
