+++ pkgs/standards/peagen/peagen/jsonschemas/extras/test5.schema.v1.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/extras/test5.schema.json",
+  "title": "EXTRAS Schema for test5",
+  "type": "object",
+  "properties": {},
+  "additionalProperties": false
+}
