+++ pkgs/standards/peagen/peagen/jsonschemas/projects_payload.schema.v1.json
@@ -1,0 +1,138 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://example.com/projects_payload.schema.json",
+  "title": "Projects Payload Schema (v1.0.0)",
+  "schemaVersion": "1.0.0",
+  "type": "object",
+  "properties": {
+    "schemaVersion": {
+      "type": "string",
+      "const": "1.0.0",
+      "description": "Projects payload schema version"
+    },
+    "PROJECTS": {
+      "type": "array",
+      "items": {
+        "oneOf": [
+          { "$ref": "#/definitions/projectGroup" }
+        ]
+      },
+      "description": "Either individual project entries with experiments or grouped payloads with META"
+    },
+    "SOURCE": {
+      "$ref": "#/definitions/source",
+      "description": "Optional metadata about spec/template and checksum"
+    }
+  },
+  "required": ["schemaVersion", "PROJECTS"],
+  "additionalProperties": false,
+  "allOf": [
+    {
+      "if": {
+        "properties": {
+          "PROJECTS": {
+            "contains": { "required": ["EXPERIMENT"] }
+          }
+        },
+        "required": ["PROJECTS"]
+      },
+      "then": { "required": ["SOURCE"] }
+    }
+  ],
+
+  "definitions": {
+    "projectGroup": {
+      "type": "object",
+      "properties": {
+        "PROJECTS": {
+          "type": "array",
+          "items": { "$ref": "#/definitions/projectDefinition" }
+        },
+        "META": {
+          "type": "object",
+          "properties": {
+            "design_id":      { "type": "string" },
+            "LLM_FACTORS":    { "type": "object", "additionalProperties": true },
+            "factors":        { "type": "object", "additionalProperties": true },
+            "spec_name":      { "type": "string" },
+            "peagen_version": {
+              "type": "string",
+              "pattern": "^(?:0|[1-9]\\d*)\\.(?:0|[1-9]\\d*)\\.(?:0|[1-9]\\d*)$"
+            }
+          },
+          "required": ["design_id", "LLM_FACTORS", "factors", "spec_name", "peagen_version"],
+          "additionalProperties": true
+        }
+      },
+      "additionalProperties": true
+    },
+
+    "package": {
+      "type": "object",
+      "properties": {
+        "NAME":                  { "type": "string" },
+        "TEMPLATE_SET_OVERRIDE": { "type": "string" },
+        "EXTRAS":                { "type": "object" },
+        "MODULES": {
+          "type": "array",
+          "items": { "$ref": "#/definitions/module" }
+        }
+      },
+      "required": ["NAME", "MODULES"],
+      "additionalProperties": true
+    },
+
+    "module": {
+      "type": "object",
+      "properties": {
+        "NAME":   { "type": "string" },
+        "EXTRAS": { "type": "object" }
+      },
+      "required": ["NAME", "EXTRAS"],
+      "additionalProperties": false
+    },
+
+    "experiment": {
+      "type": "object",
+      "properties": {
+        "FACTORS": {
+          "type": "object",
+          "patternProperties": {
+            "^[A-Za-z0-9_]+$": { "type": ["string", "number", "boolean"] }
+          },
+          "additionalProperties": false
+        }
+      },
+      "required": ["FACTORS"],
+      "additionalProperties": true
+    },
+
+    "projectDefinition": {
+      "type": "object",
+      "properties": {
+        "NAME":     { "type": "string" },
+        "ROOT":     { "type": "string" },
+        "PACKAGES": {
+          "type": "array",
+          "items": { "$ref": "#/definitions/package" }
+        }
+      },
+      "required": ["NAME", "ROOT", "PACKAGES"],
+      "additionalProperties": true
+    },
+
+    "source": {
+      "type": "object",
+      "properties": {
+        "spec":          { "type": "string" },
+        "template":      { "type": "string" },
+        "context": {
+          "type": ["string", "null"]
+        },
+        "spec_checksum": { "type": "string" }
+      },
+      "required": ["spec", "template", "spec_checksum"],
+      "additionalProperties": true
+    }
+  }
+}
