+++ scripts/generate_extras_schemas.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from typing import List
 
 TEMPLATES_ROOT = Path("pkgs/standards/peagen/peagen/templates")
-SCHEMAS_DIR = Path("pkgs/standards/peagen/peagen/schemas/extras")
+SCHEMAS_DIR = Path("pkgs/standards/peagen/peagen/jsonschemas/extras")
 
 
 def _parse_keys(md_path: Path) -> List[str]:
